Move controller to endpoints package
